🎉 Add new images to GiftBox component
This commit adds two new images to the GiftBox component. The first image is a circuit image with the alt text "indice cadeau circuit de lohéac" and dimensions of 250x150 pixels. The second image is an Alpine 110S image with the alt text "indice cadeau alpine 110S" and dimensions of 250x150 pixels.